# docs/0.5_api_contract.yaml
openapi: 3.0.3
info:
  title: OriginTrace API
  version: 0.1.0
servers:
  - url: http://localhost:8000/api
paths:
  /health:
    get:
      responses:
        '200':
          description: Service health
  /auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200': { description: JWT and public user }
        '400': { description: Missing credentials }
        '401': { description: Invalid credentials }
  /me:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current user }
        '401': { description: Unauthorized }
  /instructor/submissions:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Submission list }
        '401': { description: Unauthorized }
        '403': { description: Instructor role required }
  /student/self-checks/quota:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Daily quota }
        '401': { description: Unauthorized }
        '403': { description: Student role required }
  /analyze:
    post:
      security: [{ bearerAuth: [] }]
      description: Run the shared Git or upload analysis pipeline for an instructor submission or student self-check.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [language]
              properties:
                source_url: { type: string, format: uri }
                upload: { type: string, format: binary }
                language: { type: string, enum: [c, java, python, php] }
                is_self_check: { type: boolean, default: false }
      responses:
        '200':
          description: Analysis result with commit-history and provenance findings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '400': { description: Invalid source or language }
        '401': { description: Unauthorized }
        '403': { description: Instructor or student role required }
        '429': { description: Student daily self-check limit reached }
components:
  schemas:
    AnalysisResponse:
      type: object
      required: [submission_id, risk_band, structural_score, matches, commit_signals, provenance_flags, guidance]
      properties:
        submission_id: { type: string, format: uuid }
        risk_band: { type: string, enum: [LOW, MEDIUM, HIGH] }
        structural_score: { type: number, format: float }
        matches:
          type: array
          items:
            type: object
            required: [peer_submission_id, student_name, overlap_pct, fragments]
            properties:
              peer_submission_id: { type: string, format: uuid }
              student_name: { type: string }
              overlap_pct: { type: number, format: float }
              fragments: { type: array, items: { type: object } }
        commit_signals:
          type: object
          required: [commit_count, timespan_days, has_big_bang, low_entropy_count, author_committer_match_pct]
          properties:
            commit_count: { type: integer }
            timespan_days: { type: number }
            has_big_bang: { type: boolean }
            low_entropy_count: { type: integer }
            author_committer_match_pct: { type: number }
        provenance_flags:
          type: array
          items:
            type: object
            required: [flag_type, flag_level, description]
            properties:
              flag_type: { type: string }
              flag_level: { type: string, enum: [SOFT, HARD] }
              description: { type: string }
        guidance: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT